requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pillow==11.2.1
pymongo==4.13.2
pyperclip==1.9.0
python-dotenv==1.0.0
screeninfo==0.8.1
typing_extensions==4.14.0
ursina==7.0.0
Werkzeug==3.1.3
